Big change
This is a second  step in my avionics deal.
First of, it takes into consideration that YOUR RIG has at least 6
buttons (thats the minimum that I allow myself to work with, I hate
those minimalistic designs with 1 or 2 buttons).

It adds a GPS screen that only works inside Kerbin SOI, it also makes a
nice "ping" trail that I want to add some way to make it look a little
better without using a 128x96 array (I'll draw lines, it is written
there, but I don't know if it works, didn't had the time to test.)

Some other stuff, like finally adding the HID to control stuff like SAS,
RCS, Stage, action groups, etc, BUT there is something wrong with the
code... again, didn't had the time to test it troughtly...