Harder Stronger Better
I just made the RX buffer of the ATMega2560 a little bigger, from 64bits
to 256bits, it made the system run seamlesly. Other than that I had the
project running with an enhanced and more logical artificial horizon,
with a 3D arrow that simulate the ship pitch and roll, the horizon
itself and a heading compass over the top.
The buttons for HMI are working properly and now I can start the
different action groups trough the arduino.

Also, to point out, the TVOut is running with PAL config instead of NTSC
because the resolution and refresh is much more stable.